Change speed to int32

+++ openrtx/include/core/gps.h
@@ -51,7 +51,7 @@ typedef struct
     int32_t    latitude;             // Latitude coordinates
     int32_t    longitude;            // Longitude coordinates
     int32_t    altitude;             // Antenna altitude above mean sea level (geoid) in m
-    float      speed;                // Ground speed in km/h
+    int32_t    speed;                // Ground speed in km/h
     float      tmg_mag;              // Course over ground, degrees, magnetic
     float      tmg_true;             // Course over ground, degrees, true
 }
